Moved Colors to a separate file, updated includes
