+++ client/CMakeLists.txt
@@ -54,6 +54,7 @@ set(client_SRCS
 	render/CFadeAnimation.cpp
 	render/Canvas.cpp
 	render/ColorFilter.cpp
+	render/Colors.cpp
 	render/Graphics.cpp
 	render/IFont.cpp
 
@@ -161,6 +162,7 @@ set(client_HEADERS
 	render/CFadeAnimation.h
 	render/Canvas.h
 	render/ColorFilter.h
+	render/Colors.h
 	render/Graphics.h
 	render/ICursor.h
 	render/IFont.h
