+++ client/gui/CGuiHandler.cpp
@@ -15,6 +15,7 @@
 #include "CursorHandler.h"
 
 #include "../CGameInfo.h"
+#include "../render/Colors.h"
 #include "../renderSDL/SDL_Extensions.h"
 #include "../CMT.h"
 #include "../CPlayerInterface.h"
