+++ client/renderSDL/CursorSoftware.cpp
@@ -11,9 +11,9 @@
 #include "StdInc.h"
 #include "CursorSoftware.h"
 
-#include "SDL_Extensions.h"
-
+#include "../render/Colors.h"
 #include "../render/IImage.h"
+#include "SDL_Extensions.h"
 
 #include <SDL_render.h>
 #include <SDL_events.h>
