+++ client/widgets/Buttons.h
@@ -10,7 +10,7 @@
 #pragma once
 
 #include "../gui/CIntObject.h"
-#include "../renderSDL/SDL_Extensions.h"
+#include "../render/Colors.h"
 #include "../../lib/FunctionList.h"
 
 #include <SDL_pixels.h>
