+++ client/widgets/TextControls.cpp
@@ -17,6 +17,8 @@
 #include "../gui/CGuiHandler.h"
 #include "../windows/CMessage.h"
 #include "../adventureMap/CInGameConsole.h"
+#include "../renderSDL/SDL_Extensions.h"
+
 #include "../../lib/CGeneralTextHandler.h"
 
 #ifdef VCMI_ANDROID
