+++ client/widgets/TextControls.h
@@ -11,7 +11,7 @@
 
 #include "../gui/CIntObject.h"
 #include "../gui/TextAlignment.h"
-#include "../renderSDL/SDL_Extensions.h"
+#include "../render/Colors.h"
 #include "../render/Graphics.h"
 #include "../../lib/FunctionList.h"
